fix: optimie flash memory, add compiler flags, remove simulation mode, update wiki
