+++ platformio.ini
@@ -15,6 +15,8 @@ build_flags =
   -fno-exceptions
   -fno-threadsafe-statics
   -Os
+  -mcall-prologues
+  -fno-inline-functions-called-once
   -flto
   -ffunction-sections
   -fdata-sections
